joblogger uses 2 levels of processes instead of 3
Joblogger now only requires 2 levels of processes to start a job, the
joblogger itself and the job it started.  Previously 3 levels were
required, the joblogger, a subshell, and the job started by the
subshell.

The subshell was needed to protect the joblogger, to allow the job to
execute multiple commands, and to record any signal it catches, but it
was not very clean to start an extra process, no job in practice ran
more than one command, and the signal wouldn't necessarily be caught if
the job was killed directly without killing the joblogger or the
subshell.

This change makes the monitoring of running jobs cleaner, at the loss of
unused and unreliable features.  Note that now the PID reported in the
token is the PID of the job itself rather than the subshell which
further makes monitoring cleaner.